Read files as array buffers
readAsBinaryString should not be used because the specification states
that its result must consist of characters in the range 0..255. This, in
combination with the 16-bit unsigned integers the ecmascript
specification states that javascript strings must consist of, means that
some code points in the original data will be split into two or more code
points.
readAsArrayBuffer does not have this problem and is recommended by the
w3c specification.

https://w3c.github.io/FileAPI
https://www.ecma-international.org/ecma-262/6.0