Add login hash authorization to main view
To handle netpunkt sessions, a hash must be supplied when a user is
redirected from netpunkt to merkur. This hash is then validated by a
netpunkt service.
If this validation fails or the hash is missing, the user is supposed
to be redirected to the netpunkt login screen, but this functionality is
not implemented yet.